==> _episodes_rmd/05-multifiles.Rmd <==
---
title: "Loading multiple files"
teaching: 30
exercises: 0
questions:
- "Example ?"
objectives:
- "Example objective"
keypoints:
- "Example keypoint"
---

```{r, include = FALSE}
source("../bin/chunk-options.R")
knitr_fig_path("05-multi-R-")

```

```{r, echo=FALSE, message=FALSE}
library(tidyverse)
library(lubridate)
load("checkpoints/04end.RData")
```

## Loading more than one file at once

We have many years of weather data, and we would like to load them into a single tibble. Let's write a new function, `loadMultipleWeatherData()`, which will call our `loadWeatherData()` function, to do this.


Our function will take a vector of weather data's filenames. Before we move onto writing the function, let's think about how we will generate the vector of filenames. We could type these in by hand, for example:

```
weatherfiles <- c("data/met_mlo_insitu_1_obop_hour_1977.txt", "data/met_mlo_insitu_1_obop_hour_1978.txt", "data/met_mlo_insitu_1_obop_hour_1979.txt", ...)
```
But this would be slow and error prone.  We would also need to remember to manually update the vector every time we obtained a new year's worth of data.   Fortunately there is a function, `list.files()` which will return a vector of the files in a directory:

```{r}
list.files(path = "data")
```

We have some other files in the directory, which we don't want to try and load (such as the gapminder data).
All the files we are interested in have a consistent filename pattern: `met_mlo_insitu_1_obop_hour_yyyy.txt`, 
where `yyyy` is the year.   

The `list.files()` function takes an optional argument, `pattern`; this accepts a _regular expression_.  Unfortunately we don't have time to cover these in any depth at all.   Regular expressions provide an incredibly flexible way of matching text patterns.   They can be fiddly to write (and are often even harder to understand once you've written them).   

A regular expression uses a mixture of literal text (i.e. "normal" text), and special characters to define what matches.    If we specify part of the file that matches FIXME rewrite...

```{r}
list.files(path = "./data", pattern =  "met_mlo_insitu_1_obop_hour_")
```

`list.files()` now only returns the files we wish to load.

## More on regular expressions

Regular expressions are really useful if you are working with text.  For example, you could use them to validate that a postcode or telephone number has the correct format, or to extract parts of strings, or to perform complicated searching and replacing within strings. 

Regular expressions look intimidating, and can be fiddly to write.  If you are doing any work with text it is well worth investing the time to learn (at least the basics of) regular expressions.   A nice interactive tutorial can be found [here](https://regexone.com/)

The site [regex101](https://regex101.com/) lets you
interactively write a regular expression, and test it on a sample of data.  

You will find that your regular expressions contain a lot of `\`s.  Unfortunately R treats the
`\` as a special character in strings.  This means that we need to tell R to treat a `\` as a regular
character; we do this by _escaping_ it.  We do this with another `\`:

```{r}
mystring <- "We write a backslash as \\"
```

{: .callout}

We are almost there with listing our files.   The file names don't contain the file path, which we will need to use when we load them. 

## Challenge

Take a look at the help page for `list.files()`.  What option do you need to use to return the file path?
Use this to generate a vector, `weatherfiles`, containing the filename (with relative path) for each data file.

## Solution

The `full.names` option will return the file names with a directory path:

```{r}
weatherfiles <- list.files(path = "./data", "met_mlo_insitu_1_obop_hour_", full.names = TRUE)
```

{: .solution}
{: .challenge}

## Back to our function

We need to load each file in, and then combine them all into a single file, and then clean them.

One approach to this might be to load each file in, and then append it to a master data-set.   This will work, but is actually quite slow, since the R will have to take a copy of the entire master data set each time we add more data to it.

A more efficient way of performing the task is to make a list of tibbles, and then combine these in a single step after we've loaded all of the data.

## Lists

We have previously encountered vectors, which we can create using the `c()` function:

```{r}
c(1,2,4,8)
c("a", "b", "c")
```

All of the elements of a vector must be the same data type; R will make this so if we try and make a vector with different data types:

```{r}
c(1, "a", 2, 3)
```

We can think of a _list_ as a generalisation of this; each element of the list can contain pretty much any other object.  We make a list with the `list()` function:

```{r}
mylist <- list(1, "a", 2, c(1,2,3))
mylist
```
This has created a list; the first element contains the number 1, the second the letter "a", the third contains the number 3 and the 4th element contains a vector, which contains the values 1,2 and 3.

We can use the subsetting operator, `[]` on the list in the same way we would a vector. This will return another list:

```{r}
mylist[2]
mylist[2:3]

```

To refer to the contents of a list element we use the `[[]]` operator.  In contrast to the `[]` operator, this takes a single value:

```{r}
mylist[[1]]
mylist[[4]]
```
But note we cannot use, e.g.:

```{r, eval=FALSE}
mylist[[1:2]]
```
Since this would return more than one element.

There is another way of manipulating lists, which is to name its elements:

```{r}
weatherlist <- list(temperature = 2.5, windspeed = 4, winddir = "N")
```

We can still refer to the list elements by number
```{r}
weatherlist[1:2]
```
We can also refer to them by name:

```{r}
# This will return a list
weatherlist[c("windspeed", "winddir")]
# As will this
weatherlist[c("windspeed")]

## We use [[]] to access the data itself:
weatherlist[["windspeed"]]
```

We can also return the contents of a list element with the `$` operator:

```{r}
weatherlist$windspeed
```

One useful feature of the `[[]]` operator is that we can define the element we want in a variable:

```{r}
myvar <- "windspeed"
weatherlist[[myvar]]
```

We cannot do this with the `$` operator.

We can add or update an element of a list using the `[[]]` or `$` operators:

```{r}
weatherlist$windspeed <- 3
weatherlist[["winddir"]] <- "E"
```

You can only add elements to a list if it already exists.  You can create a new (empty) list using the `list()` function with no arguments.

To delete an element from a list assign the element the value `NULL`, e.g. `weatherlist$deleteme <- NULL`.

## Loading our files

This is what we want our function to do:

```
loadWeatherDataPseudoCode <- function(weatherfiles){
  Make an empty list
  for i in 1:length(i) {
    Load in data for file weatherfiles[i]
    Append the data to the ith element of the list
  }
  convert the list into a single tibble
  
}

loadWeatherDataPseudoCode(weatherfiles)

```


## Challenge 

Taking the pseudocode above as a starting point, create a function to load more than one weather file.
The function should return a list, each element of which contains a tibble of weather data (we will cover
converting the list into a single tibble after the challenge).

## Solution

```{r}

loadMultipleWeatherData <- function(weatherfiles){
  weatherList <- list()
  for (i in 1:length(weatherfiles)) {
    weatherList[[i]] <- loadWeatherData(weatherfiles[i])
  }
  
  return(weatherList)
}

# Load in a couple of years data:
weatherDataList <- loadMultipleWeatherData(weatherfiles[1:2])

```

{: .solution}
{: .challenge}


## Generating a single tibble

We use the `bind_rows()` function to combine multiple tibbles into a single tibble.  We can pass a list of tibbles as the argument to the function, or provide one or more tibbles that we wish to concatenate.

Let's modify our function to convert the list of tibbles into a single tibble:

```{r}

loadMultipleWeatherData <- function(weatherfiles){
  weatherList <- list()
  for (i in 1:length(weatherfiles)) {
    weatherList[[i]] <- loadWeatherData(weatherfiles[i])
  }

  # Convert list of tibbles to a single tibble
  weatherData <- bind_rows(weatherList) 
  return(weatherData)
}

# Load in a couple of years data:
weatherData <- loadMultipleWeatherData(weatherfiles[1:2])

```

Let's load in the data for all years:

```{r}
cleanweather <- loadMultipleWeatherData(weatherfiles)
```

We can sense check our load by checking that the number of observations in each year is (roughly - don't forget leap years) the same.

```{r}

cleanweather %>%  
  group_by(yyyy) %>% 
  count()  
```

And by checking the summary statistics for any unusual values:

```{r}

summary(cleanweather)
```

## Summary

In this episode we've written a function to load in more than 1 weather data file at once, and to return a tibble containing all of the loaded data.   


```{r, echo=FALSE}
save.image(file = "checkpoints/05end.RData")
```
